Initial version of the sync script.
This synchronizes (uncommitted) changes with a remote repository.

The sync is performed using a non-head ref under the 'refs/sync' namespace.

This effectively works as a distributed undo/redo history stored in
the git repository.

Currently, that history is never reset. A later change will modify
this so that the history is reset every time the latest changes
are committed.